PHP Divisor: Combine statementsto get rid of braces (72 -> 71).
By combining both statements in the inner while loop with a logical operator, one can get rid of the curly braces. Since logical operators are two characters instead of one for the semicolon, this only works for the case of two inner statements, thus not for the outer while loop.